slowcontrol/requirements.txt
# xsphere slow control service dependencies
# Install with: pip install -r requirements.txt

# MQTT
paho-mqtt>=2.0.0

# Modbus TCP (PLC)
pymodbus>=3.6.0

# Serial (Omega RDXL6SD-USB, future: scale, pumping stands)
pyserial>=3.5

# Configuration
pyyaml>=6.0

# Utilities
numpy>=1.24.0
